requirements.txt
flask
pyqt5
pyjwt
requests 